ExprParser

Description

Depend : This module depends on the ExprType module

Documentation

parseExprD :: String -> Expr Double

This parser takes binary construcotrs( +, *, **) and parses strings numbers into the Double type of an Expr constructor

parseExprDnlsc :: String -> Expr Double

This parser takes unary constructors(natrual log, natural exponent, sin, cos) and parses strings numbers into the Double type of an Expr constructor

parseExprF :: String -> Expr Float

This parser takes binary construcotrs( +, *, **) and parses strings numbers into the Float type of an Expr constructor

parseExprFnlsc :: String -> Expr Float

This parser takes unary constructors(natrual log, natural exponent, sin, cos) and parses strings numbers into the Float type of an Expr constructor