ExprTest

Description

Documentation

evalCase1 :: Double -> Double -> Bool

Test case for addition

evalCase2 :: Double -> Double -> Bool

Test case for multiplication

evalCase3 :: Double -> Double -> Bool

Test case for Exponent

simplifyCase1 :: Double -> Double -> Bool

Test case to check simplification of multiplication of a variable with constant

simplifyCase2 :: Double -> Double -> Bool

Test case to check exponent

diffCase :: String -> Double -> Bool

Test case to check partial differentiation of a constant